Updates April 1, 2016
Still reworking the mockup after discussing with you the structure of the page using the nav, main and section tags. Putting the nav and main tags side by side and move the id or class div's inside the section tags. I have the html page rebuilt. I will need to modify the css to fit the new structure.